move NumericalGradientCalculator to separate file
